Reorganise `decode` and `reader`
