+++ avro/Cargo.toml
@@ -60,6 +60,7 @@ apache-avro-derive = { default-features = false, version = "0.20.0", path = "../
 bigdecimal = { default-features = false, version = "0.4.9", features = ["std", "serde"] }
 bon = { default-features = false, version = "3.8.1" }
 bzip2 = { version = "0.6.1", optional = true }
+corosensei = { version = "0.3.1" }
 crc32fast = { default-features = false, version = "1.5.0", optional = true }
 digest = { default-features = false, version = "0.10.7", features = ["core-api"] }
 miniz_oxide = { default-features = false, version = "0.8.9", features = ["with-alloc"] }
