+++ avro/src/decode/datum.rs
@@ -1,8 +1,8 @@
 use oval::Buffer;
 
-use super::StateMachineResult;
-use crate::state_machines::reading::{
-    CommandTape, ItemRead, StateMachine, StateMachineControlFlow, SubStateMachine,
+use crate::decode::{
+    CommandTape, ItemRead, StateMachine, StateMachineControlFlow, StateMachineResult,
+    SubStateMachine,
 };
 
 enum TapeOrFsm {
