+++ avro/src/decode/error.rs
@@ -1,4 +1,4 @@
-use crate::{Schema, state_machines::reading::ItemRead};
+use crate::{Schema, decode::ItemRead};
 use thiserror::Error;
 
 #[derive(Error, Debug)]
