--- avro/src/state_machines/mod.rs
@@ -1,1 +1,0 @@
-pub mod reading;
